pyproject.toml
[project]
name = "replaxy"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
    "livekit-agents[mcp,silero,turn-detector]>=1.3.11",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit>=1.0.22", # 1.0.21 is yanked (bug in AudioStream); constraint forces a valid version
    "mcp",
    "mem0ai",
    "python-dotenv",
    "pyyaml",
    "rich",
    "typer",
    "tzdata",
]

[project.scripts]
replaxy = "src.cli:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
